Combine frames: move code to feature/logs
